@@ -19,53 +19,53 @@
   },
   "dependencies": {
     "autobindr": "^1.0.6",
-    "bootstrap": "^4.0.0",
-    "classnames": "^2.2.5",
+    "bootstrap": "^4.1.1",
+    "classnames": "^2.2.6",
     "formik": "^0.11.11",
     "mdi": "^2.2.43",
     "prop-types": "^15.6.1",
     "ramda": "^0.25.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-dropzone": "^4.2.9",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-dropzone": "^4.2.11",
     "react-redux": "^5.0.7",
     "react-router": "3",
     "react-router-dom": "^4.2.2",
-    "react-toastify": "^4.0.1",
-    "reactstrap": "^5.0.0-beta.3",
+    "react-toastify": "^4.1.0",
+    "reactstrap": "^6.1.0",
     "recompact": "^3.3.0",
-    "redux": "^3.7.2",
+    "redux": "^4.0.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
+    "redux-thunk": "^2.3.0",
     "reselect": "^3.0.1",
     "uuid": "^3.2.1",
     "web3": "0.20.6"
   },
   "devDependencies": {
-    "babel-jest": "^22.4.3",
+    "babel-jest": "^23.0.1",
     "babel-loader": "^7.1.4",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "cross-env": "^5.1.4",
+    "cross-env": "^5.1.6",
     "css-loader": "^0.28.11",
-    "html-webpack-plugin": "^3.0.6",
+    "html-webpack-plugin": "^3.2.0",
     "http-server": "^0.11.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^22.4.3",
-    "jest-fetch-mock": "^1.5.0",
-    "mini-css-extract-plugin": "^0.2.0",
+    "jest": "^23.1.0",
+    "jest-fetch-mock": "^1.6.3",
+    "mini-css-extract-plugin": "^0.4.0",
     "ncp": "^2.0.0",
-    "node-sass": "^4.7.2",
-    "npm-run-all": "^4.1.2",
-    "prettier": "^1.11.1",
+    "node-sass": "^4.9.0",
+    "npm-run-all": "^4.1.3",
+    "prettier": "^1.13.4",
     "rimraf": "^2.6.2",
-    "sass-loader": "^6.0.7",
-    "style-loader": "^0.20.3",
-    "webpack": "^4.1.1",
-    "webpack-bundle-analyzer": "^2.11.1",
-    "webpack-cli": "^2.0.12",
-    "webpack-dev-server": "^3.1.1",
+    "sass-loader": "^7.0.3",
+    "style-loader": "^0.21.0",
+    "webpack": "^4.12.0",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^3.0.3",
+    "webpack-dev-server": "^3.1.4",
     "webpack-merge": "^4.1.2"
   }
 }
