@@ -38,7 +38,7 @@ class OutcomingExchanges extends React.Component {
   }
   // replace in React v17
   // static getDerivedStateFromProps(nextProps, prevState)
-  componentWillUpdate(nextProps) {
+  UNSAFE_componentWillUpdate(nextProps) {
     refetchExchages(nextProps);
   }
 
@@ -140,6 +140,9 @@ const mapDispatchToProps = (dispatch, ownProps) => ({
 export default compose(
   withRouter,
   withPage(DEFAULT_LIMIT),
-  connect(mapStateToProps, mapDispatchToProps),
+  connect(
+    mapStateToProps,
+    mapDispatchToProps
+  ),
   withValidPageEnsurance(({count}) => count, DEFAULT_LIMIT)
 )(OutcomingExchanges);
